created calendar endpoint, moved userInfoForm.tsx to here from signup, moved the calendar auth function in signup, setup page.tsx with similar structure as page.tsx in signup
